Fixes oada-ref-auth in oada-api-server
oada-api-server apparently assumes everyone wants a raw body parser. Quick fix
is to move the OAuth 2.0 stuff to be above the raw body parser so oada-ref-auth
short circuits requests before the damage is done.

The real solution is to apply the raw body parser to just the routes that need
it.